pyproject.toml
[project]
name = "llm-penandpaper"
version = "0.1.0"
description = "An immersive AI-powered tabletop RPG adventure with AMD GPU support, optimized code, and enhanced performance."
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "accelerate==0.25.0",
    "aiohttp>=3.10,<4",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==23.2.0",
    "banks==2.2.0",
    "blinker==1.9.0",
    "cachetools==6.1.0",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "diffusers==0.25.1",
    "dirtyjson==1.0.8",
    "distro==1.9.0",
    "edge-tts>=7,<8",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "greenlet==3.2.3",
    "griffe==1.8.0",
    "h11==0.16.0",
    "hf-xet==1.1.5",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.20.3",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.72",
    "langchain-openai==0.2.14",
    "langchain-ollama==0.3.6",
    "langchain-text-splitters==0.3.9",
    "langgraph==0.5.4",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.5.2",
    "langgraph-sdk==0.1.74",
    "langsmith==0.4.8",
    "langfuse>=2,<3",
    "llama-index-core==0.12.52.post1",
    "llama-index-embeddings-huggingface==0.5.5",
    "llama-index-instrumentation==0.3.0",
    "llama-index-llms-langchain==0.6.1",
    "llama-index-readers-file==0.4.11",
    "llama-index-workflows==1.2.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mpmath==1.3.0",
    "multidict==6.6.3",
    "mypy-extensions==1.1.0",
    "narwhals==1.48.1",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "nltk==3.9.1",
    "numpy==1.26.4",
    "ollama==0.5.1",
    "openai==1.97.1",
    "litellm[proxy]>=1.72,<2",
    "orjson==3.11.1",
    "ormsgpack==1.10.0",
    "packaging==23.2",
    "pandas==1.5.3",
    "pillow==10.2.0",
    "pip==25.1.1",
    "platformdirs==4.3.8",
    "propcache==0.3.2",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "pyarrow==21.0.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydeck==0.9.1",
    "pypdf==5.8.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.31.0",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.26.0",
    "safetensors==0.5.3",
    "scikit-learn==1.7.1",
    "scipy==1.15.3",
    "sentence-transformers==3.1.1",
    "setuptools==80.9.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.41",
    "streamlit==1.47.1",
    "striprtf==0.0.26",
    "sympy==1.14.0",
    "tenacity==8.5.0",
    "threadpoolctl==3.6.0",
    "tiktoken==0.9.0",
    "tokenizers==0.15.0",
    "toml==0.10.2",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "transformers==4.38.0",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wrapt==1.17.2",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.23.0",
]

[project.optional-dependencies]
amd-gpu = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torchaudio>=2.1.0",
    "xformers>=0.0.22",
    "flash-attn>=2.3.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
